backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pyserial==3.5
pydantic==2.9.2
python-multipart==0.0.12

